requirements.txt
python-magic==0.4.15
Pillow==6.1.0
flask==1.1.1
connexion==2.3.0
swagger-ui-bundle
gevent